When the Today, Yesterday, This Week, Last Week, This Month, Last Month, This Year, Last Year buttons are selected the resoultion is set for mins for Today, Yesterday, This Week, Last Week and to Hours for This Month, Last Month, This Year, Last Year. This allows for a good trade off betweeen plot time and plot resolution.
Increase the default maximum number of plot points that will be plotted from 10000 to 11000 so that a weeks with of min resolution data can be plotted.

If the user wants to plot many plot points over a longer period they can still select the Second, Min, Hour or Day button and then select the power button. They may need to use the -m command line option to set the max imum number of plot points.