Further to the initial work on a single app that
    - Detects CT6 units on the LAN/WiFi.
    - Regularly (every second) reasdin stats from CT6 these CT6 units.
    - Saves this data to sqlite3 databases.
    - Presents a user friendly web interface to view and manipulate the data.

This commit includes the last of the above, a GUI (same as ct6_dash.py) that
reads data from local sqlite databases rather than the mysql databases used
when ct6_db_store.py is used. This simplifies the software required as rather
than having a mysql server with the ct6_db_store.py feeding it data and ct6_dash
providing the web GUI to allow the user to interact with the data, all that
is now required is to run the ct6_app.py.

Another advantage of this approach is that larger datasets (E.G 24 hours of
second sampled data, 86400 samples) can be plotted in half the time
compared to when a mysql database was used.